fix(mediarequest): move methods modifying MediaRequest to its Subscriber
